id 관련 필드 Long타입으로 변경 작업

+++ backend/src/main/java/com/bada/cali/dto/LogDTO.java
@@ -12,15 +12,15 @@ import java.time.LocalDateTime;
 @Builder
 public class LogDTO {
 	
-	private Integer id;              // LOG 고유번호
+	private Long id;              // LOG 고유번호
 	private String logIp;            // LOG ip
 	private String workerName;       // 작업자 명
 	private String logContent;       // LOG 이력내용
 	private String logType;          // 로그타입 (i:등록, u:수정, d:삭제, l:로그인)
 	private String refTable;         // 참조 테이블명
-	private Integer refTableId;      // 참조 테이블 idx
+	private Long refTableId;      // 참조 테이블 idx
 	private LocalDateTime createDatetime; // 등록일시
-	private Integer createMemberId;  // 등록자
+	private Long createMemberId;  // 등록자
 	
 	// 엔티티 → DTO 변환 메서드
 	public static LogDTO fromEntity(Log log) {
